Revert "[CLI] move out localnet logic + add command for Aptos Workspace (#15508)"
This reverts commit 803b7fda81ab9f17fa82b630fcd50954cee25039.
