+++ aptos-move/aptos-workspace-server/Cargo.toml
@@ -13,10 +13,10 @@ rust-version = { workspace = true }
 
 [dependencies]
 # aptos deps
+aptos = { workspace = true }
 aptos-cached-packages = { workspace = true }
 aptos-config = { workspace = true }
 aptos-faucet-core = { workspace = true }
-aptos-localnet = { workspace = true }
 aptos-node = { workspace = true }
 aptos-types = { workspace = true }
