+++ aptos-move/aptos-workspace-server/src/services/node.rs
@@ -3,8 +3,8 @@
 
 use crate::common::IP_LOCAL_HOST;
 use anyhow::{bail, Result};
+use aptos::node::local_testnet::HealthChecker;
 use aptos_config::config::{NodeConfig, TableInfoServiceMode};
-use aptos_localnet::health_checker::HealthChecker;
 use aptos_node::{load_node_config, start_and_report_ports};
 use aptos_types::network_address::{NetworkAddress, Protocol};
 use futures::channel::oneshot;
