--- crates/aptos-localnet/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "aptos-localnet"
-version = "0.1.0"
-
-# Workspace inherited keys
-authors = { workspace = true }
-edition = { workspace = true }
-homepage = { workspace = true }
-license = { workspace = true }
-publish = { workspace = true }
-repository = { workspace = true }
-rust-version = { workspace = true }
-
-[dependencies]
-anyhow = { workspace = true }
-aptos-indexer-grpc-utils = { workspace = true }
-bollard = { workspace = true }
-dirs = { workspace = true }
-futures = { workspace = true }
-tracing = { workspace = true }
-version-compare = { workspace = true }
-
-aptos-protos = { workspace = true }
-aptos-rest-client = { workspace = true }
-diesel = { workspace = true, features = [
-    "postgres_backend",
-] }
-diesel-async = { workspace = true }
-processor = { git = "https://github.com/aptos-labs/aptos-indexer-processors.git", rev = "51a34901b40d7f75767ac907b4d2478104d6a515", default-features = false }
-reqwest = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-tonic = { workspace = true }
