+++ crates/aptos/src/node/local_testnet/processors.rs
@@ -3,15 +3,19 @@
 
 use super::{health_checker::HealthChecker, traits::ServiceManager, RunLocalnet};
 use anyhow::{bail, Context, Result};
-pub use aptos_localnet::processors::get_processor_config;
 use async_trait::async_trait;
 use clap::Parser;
 use diesel::Connection;
 use diesel_async::{async_connection_wrapper::AsyncConnectionWrapper, pg::AsyncPgConnection};
 use maplit::hashset;
 use processor::{
-    gap_detectors::DEFAULT_GAP_DETECTION_BATCH_SIZE, processors::ProcessorName,
-    utils::database::run_pending_migrations, IndexerGrpcProcessorConfig,
+    gap_detectors::DEFAULT_GAP_DETECTION_BATCH_SIZE,
+    processors::{
+        objects_processor::ObjectsProcessorConfig, stake_processor::StakeProcessorConfig,
+        token_v2_processor::TokenV2ProcessorConfig, ProcessorConfig, ProcessorName,
+    },
+    utils::database::run_pending_migrations,
+    IndexerGrpcProcessorConfig,
 };
 use reqwest::Url;
 use server_framework::RunnableConfig;
@@ -51,6 +55,64 @@ pub struct ProcessorArgs {
 pub struct ProcessorManager {
     config: IndexerGrpcProcessorConfig,
     prerequisite_health_checkers: HashSet<HealthChecker>,
+}
+
+pub fn get_processor_config(processor_name: &ProcessorName) -> Result<ProcessorConfig> {
+    Ok(match processor_name {
+        ProcessorName::AccountTransactionsProcessor => {
+            ProcessorConfig::AccountTransactionsProcessor
+        },
+        ProcessorName::AnsProcessor => {
+            bail!("ANS processor is not supported in the localnet")
+        },
+        ProcessorName::DefaultProcessor => ProcessorConfig::DefaultProcessor,
+        ProcessorName::EventsProcessor => ProcessorConfig::EventsProcessor,
+        ProcessorName::FungibleAssetProcessor => ProcessorConfig::FungibleAssetProcessor,
+        ProcessorName::MonitoringProcessor => {
+            bail!("Monitoring processor is not supported in the localnet")
+        },
+        ProcessorName::NftMetadataProcessor => {
+            bail!("NFT Metadata processor is not supported in the localnet")
+        },
+        ProcessorName::ObjectsProcessor => {
+            ProcessorConfig::ObjectsProcessor(ObjectsProcessorConfig {
+                query_retries: Default::default(),
+                query_retry_delay_ms: Default::default(),
+            })
+        },
+        ProcessorName::ParquetDefaultProcessor => {
+            bail!("ParquetDefaultProcessor is not supported in the localnet")
+        },
+        ProcessorName::ParquetFungibleAssetProcessor => {
+            bail!("ParquetFungibleAssetProcessor is not supported in the localnet")
+        },
+        ProcessorName::ParquetTransactionMetadataProcessor => {
+            bail!("ParquetTransactionMetadataProcessor is not supported in the localnet")
+        },
+        ProcessorName::ParquetAnsProcessor => {
+            bail!("ParquetAnsProcessor is not supported in the localnet")
+        },
+        ProcessorName::ParquetEventsProcessor => {
+            bail!("ParquetEventsProcessor is not supported in the localnet")
+        },
+        ProcessorName::ParquetTokenV2Processor => {
+            bail!("ParquetTokenV2Processor is not supported in the localnet")
+        },
+        ProcessorName::StakeProcessor => ProcessorConfig::StakeProcessor(StakeProcessorConfig {
+            query_retries: Default::default(),
+            query_retry_delay_ms: Default::default(),
+        }),
+        ProcessorName::TokenV2Processor => {
+            ProcessorConfig::TokenV2Processor(TokenV2ProcessorConfig {
+                query_retries: Default::default(),
+                query_retry_delay_ms: Default::default(),
+            })
+        },
+        ProcessorName::TransactionMetadataProcessor => {
+            ProcessorConfig::TransactionMetadataProcessor
+        },
+        ProcessorName::UserTransactionProcessor => ProcessorConfig::UserTransactionProcessor,
+    })
 }
 
 impl ProcessorManager {
